TBC tool warnings
Many warnings in the TBC tools have been resolved.

+++ src/tools/ToolsTBC/mmaps_generator/PathGenerator.cpp
@@ -85,7 +85,7 @@ bool handleArgs(int argc, char** argv,
             if (!param)
                 return false;
 
-            float maxangle = atof(param);
+            float maxangle = static_cast<float>(atof(param));
             if (maxangle <= 90.f && maxangle >= 45.f)
                 maxAngle = maxangle;
             else
@@ -293,6 +293,6 @@ int main(int argc, char** argv)
         builder.buildAllMaps(threads);
 
     if (!silent)
-        printf("Finished. MMAPS were built in %u ms!\n", Util::GetTimeDifferenceToNow(startTime));
+        printf("Finished. MMAPS were built in %lld ms!\n", Util::GetTimeDifferenceToNow(startTime));
     return 0;
 }
